refactoring: controllers match roles
