+++ main/src/main/kotlin/com/ip13/main/controller/ManagerController.kt
@@ -1,13 +1,14 @@
 package com.ip13.main.controller
 
 import com.ip13.main.model.dto.request.AddBookingConstraintRequest
+import com.ip13.main.model.dto.request.GradeManagerRequest
 import com.ip13.main.model.dto.request.ReservationProcessRequest
-import com.ip13.main.model.dto.request.TableReserveRequest
 import com.ip13.main.model.dto.response.AddBookingConstraintResponse
+import com.ip13.main.model.dto.response.GradeManagerResponse
 import com.ip13.main.model.dto.response.ReservationProcessResponse
 import com.ip13.main.model.dto.response.ShowReservationsResponse
-import com.ip13.main.model.dto.response.TableReserveResponse
 import com.ip13.main.service.BookingConstraintService
+import com.ip13.main.service.GradeManagerService
 import com.ip13.main.service.TableReserveService
 import com.ip13.main.util.getLogger
 import jakarta.validation.Valid
@@ -17,25 +18,13 @@ import org.springframework.web.bind.annotation.*
 
 @Validated
 @RestController
-@RequestMapping("/reserve", method = [RequestMethod.POST, RequestMethod.GET])
-class ReserveController(
+@RequestMapping("/manager", method = [RequestMethod.POST, RequestMethod.GET])
+class ManagerController(
     private val tableReserveService: TableReserveService,
     private val bookingConstraintService: BookingConstraintService,
+    private val gradeManagerService: GradeManagerService,
 ) {
     private val log = getLogger(javaClass)
-
-    @PostMapping("/reserve-table")
-    fun reserveTable(
-        @RequestHeader(name = "Authorization", required = true)
-        authHeader: String,
-        @Valid
-        @RequestBody(required = true)
-        request: TableReserveRequest
-    ): TableReserveResponse {
-        log.debug("/reserve/reserve-table endpoint invoked")
-
-        return tableReserveService.reserveTable(request, authHeader)
-    }
 
     @PostMapping("/add-booking-constraint")
     fun addBookingConstraint(
@@ -77,5 +66,18 @@ class ReserveController(
         log.debug("/reserve/process-reservation endpoint invoked")
 
         return tableReserveService.processReservation(authHeader, request)
+    }
+
+    @GetMapping("/add-grade-manager")
+    fun addGradeManager(
+        @RequestHeader(name = "Authorization", required = true)
+        authHeader: String,
+        @Valid
+        @RequestBody(required = true)
+        request: GradeManagerRequest,
+    ): GradeManagerResponse {
+        log.debug("/restaurant/add_grade-manager endpoint invoked")
+
+        return gradeManagerService.gradeUser(authHeader, request)
     }
 }
